removed unneeded demo integration code
